Parse the ChangeLog instead of a Maildir
Not huge changes, but important, means it is less likely to blow up if I don't
clean out my mailbox and use less memory.

Also includes some improvements in the parsing of the commit message, so
should be fewer "did something the parser couldn't recognize" messages.

Plus a script to sync the changelog